# atlas-core/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/login:
        post:
            tags:
                - Auth
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.LoginReply'
    /v1/logout:
        post:
            tags:
                - Auth
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.LogoutReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.LogoutReply'
...
    /v1/register:
        post:
            tags:
                - Auth
            operationId: Auth_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.RegisterReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.RegisterReply'
...
components:
    schemas:
...
        api.core.v1.LoginReply:
            type: object
            properties:
                token:
                    type: string
        api.core.v1.LoginReq:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                uuid:
                    type: string
                code:
                    type: string
        api.core.v1.LogoutReply:
            type: object
            properties: {}
        api.core.v1.LogoutReq:
            type: object
            properties: {}
...
        api.core.v1.RegisterReply:
            type: object
            properties: {}
        api.core.v1.RegisterReq:
            type: object
            properties:
                username:
                    type: string
                nickname:
                    type: string
                mobile:
                    type: string
                password:
                    type: string
...